src/test/java/com/mystore/testcases/IndexPageTest.java
/**
 * 
 */
package com.mystore.testcases;

import org.testng.Assert;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Parameters;
import org.testng.annotations.Test;

import com.mystore.base.BaseClass;
import com.mystore.pageobjects.IndexPage;

/**
 * 
 */
public class IndexPageTest extends BaseClass{
	IndexPage indexPage;
	
	@Parameters("browser")
	@BeforeMethod(groups = {"Smoke", "Sanity", "Regression"})
	public void setup(String browser) {
		launchApp(browser);
	}
	
	@AfterMethod(groups = {"Smoke", "Sanity", "Regression"})
	public void tearDown()  {
		getDriver().quit();
	}
	
	@Test(groups = "Smoke" )
	public void verifyLogo() throws Throwable {
		indexPage = new IndexPage();
		boolean resultLogo = indexPage.validateLogo();
		Assert.assertTrue(resultLogo);
	}
	
	
}
